Add create MusicAlbumsManager, GamesManager.rb, and BooksManager classes
