+++ books.json
@@ -1,1 +1,1 @@
-[{"id":818,"publish_date":"2012-12-01","publisher":"Noel","cover_state":"good"},{"id":60,"publish_date":"2013-12-01","publisher":"Pamy","cover_state":"bad"},{"id":234,"publish_date":"2012-12-12","publisher":"wasaa","cover_state":"badd"},{"id":731,"publish_date":"2019-12-01","publisher":"Noel","cover_state":"good"},{"id":527,"publish_date":"2016-12-01","publisher":"Pam","cover_state":"good"},{"id":5,"publish_date":"2013-11-01","publisher":"Queen","cover_state":"good"},{"id":879,"publish_date":"2018-12-01","publisher":"filbert","cover_state":"good"},{"id":628,"publish_date":"2012-03-01","publisher":"lion","cover_state":"good"},{"id":215,"publish_date":"2015-12-01","publisher":"Son","cover_state":"good"},{"id":467,"publish_date":"2016-12-12","publisher":"ryan","cover_state":"good"},{"id":263,"publish_date":"2012-12-01","publisher":"eer","cover_state":"good"},{"id":433,"publish_date":"2012-12-01","publisher":"filbrry","cover_state":"good"},{"id":530,"publish_date":"2012-12-01","publisher":"Kasee","cover_state":"bad"},{"id":949,"publish_date":"2023-12-01","publisher":"Pam","cover_state":"Good"},{"id":448,"publish_date":"2012-12-01","publisher":"goow","cover_state":"good"},{"id":540,"publish_date":"2012-12-12","publisher":"Noelso","cover_state":"good"},{"id":434,"publish_date":"2010-12-01","publisher":"Maino","cover_state":"bad"},{"id":347,"publish_date":"2012-12-01","publisher":"Pam","cover_state":"bad"},{"id":891,"publish_date":"2013-02-01","publisher":"Judy","cover_state":"good"},{"id":381,"publish_date":"2013-12-01","publisher":"King","cover_state":"bad"},{"id":579,"publish_date":"2013-12-01","publisher":"Noelo","cover_state":"Good"},{"id":690,"publish_date":"2011-11-01","publisher":"Pam","cover_state":"bad"},{"id":362,"publish_date":"2013-02-01","publisher":"Pam","cover_state":"good"},{"id":801,"publish_date":"2012-03-04","publisher":"Noelso","cover_state":"Bad"},{"id":877,"publish_date":"2019-04-01","publisher":"Kasee","cover_state":"bad"},{"id":197,"publish_date":"2012-12-01","publisher":"noel","cover_state":"good"},{"id":677,"publish_date":"2011-11-01","publisher":"Pam","cover_state":"good"},{"id":659,"publish_date":"2012-12-01","publisher":"Noel","cover_state":"good"},{"id":496,"publish_date":"2012-12-12","publisher":"Noel","cover_state":"good"},{"id":321,"publish_date":"2012-12-01","publisher":"Shirley","cover_state":"good"},{"id":670,"publish_date":"2015-12-01","publisher":"Kasee","cover_state":"bad"},{"id":490,"publish_date":"2012-12-21","publisher":"sds","cover_state":"sd"}]
+[{"id":348,"publish_date":"2012-12-01","publisher":"Noel","cover_state":"good"},{"id":409,"publish_date":"2013-12-01","publisher":"Pamy","cover_state":"bad"},{"id":340,"publish_date":"2012-12-12","publisher":"wasaa","cover_state":"badd"},{"id":5,"publish_date":"2019-12-01","publisher":"Noel","cover_state":"good"},{"id":833,"publish_date":"2016-12-01","publisher":"Pam","cover_state":"good"},{"id":791,"publish_date":"2013-11-01","publisher":"Queen","cover_state":"good"},{"id":636,"publish_date":"2018-12-01","publisher":"filbert","cover_state":"good"},{"id":607,"publish_date":"2012-03-01","publisher":"lion","cover_state":"good"},{"id":272,"publish_date":"2015-12-01","publisher":"Son","cover_state":"good"},{"id":161,"publish_date":"2016-12-12","publisher":"ryan","cover_state":"good"},{"id":523,"publish_date":"2012-12-01","publisher":"eer","cover_state":"good"},{"id":957,"publish_date":"2012-12-01","publisher":"filbrry","cover_state":"good"},{"id":280,"publish_date":"2012-12-01","publisher":"Kasee","cover_state":"bad"},{"id":64,"publish_date":"2023-12-01","publisher":"Pam","cover_state":"Good"},{"id":597,"publish_date":"2012-12-01","publisher":"goow","cover_state":"good"},{"id":335,"publish_date":"2012-12-12","publisher":"Noelso","cover_state":"good"},{"id":404,"publish_date":"2010-12-01","publisher":"Maino","cover_state":"bad"},{"id":930,"publish_date":"2012-12-01","publisher":"Pam","cover_state":"bad"},{"id":576,"publish_date":"2013-02-01","publisher":"Judy","cover_state":"good"},{"id":843,"publish_date":"2013-12-01","publisher":"King","cover_state":"bad"},{"id":432,"publish_date":"2013-12-01","publisher":"Noelo","cover_state":"Good"},{"id":6,"publish_date":"2011-11-01","publisher":"Pam","cover_state":"bad"},{"id":245,"publish_date":"2013-02-01","publisher":"Pam","cover_state":"good"},{"id":265,"publish_date":"2012-03-04","publisher":"Noelso","cover_state":"Bad"},{"id":802,"publish_date":"2019-04-01","publisher":"Kasee","cover_state":"bad"},{"id":530,"publish_date":"2012-12-01","publisher":"noel","cover_state":"good"},{"id":322,"publish_date":"2011-11-01","publisher":"Pam","cover_state":"good"},{"id":870,"publish_date":"2012-12-01","publisher":"Noel","cover_state":"good"},{"id":122,"publish_date":"2012-12-12","publisher":"Noel","cover_state":"good"},{"id":152,"publish_date":"2012-12-01","publisher":"Shirley","cover_state":"good"},{"id":784,"publish_date":"2015-12-01","publisher":"Kasee","cover_state":"bad"},{"id":158,"publish_date":"2012-12-21","publisher":"sds","cover_state":"sd"},{"id":8,"publish_date":"2021-12-21","publisher":"ddd","cover_state":"d"}]
