+++ labels.json
@@ -1,1 +1,1 @@
-[{"title":"git","color":"rd"},{"title":"gift","color":"red yeloow"},{"title":"new","color":"brown"},{"title":"adsd","color":"asd"},{"title":"sdcsd","color":"asdsad"},{"title":"sd","color":"sd"}]
+[{"title":"git","color":"rd"},{"title":"gift","color":"red yeloow"},{"title":"new","color":"brown"},{"title":"adsd","color":"asd"},{"title":"sdcsd","color":"asdsad"},{"title":"sd","color":"sd"},{"title":"wew","color":"we"}]
